refactor(ui): centralize layout and bottom navigation in App root
- Move BottomNavigation and main layout container to App root in
- Remove redundant layout wrappers and BottomNavigation from route components (, , )
- Simplify and unify responsive width logic
- Adjust styles in  and minor tweaks in BottomNavigation and TopBar for consistency
