+++ src/app.css
@@ -22,18 +22,16 @@
   /* @media (prefers-color-scheme: dark) { */
   :root {
     --foreground-rgb: 255, 255, 255;
-    --background-start-rgb: 0, 0, 0;
-    --background-end-rgb: 0, 0, 0;
   }
 
   /* } */
 
+  html {
+    background: black;
+  }
+
   body {
     color: rgb(var(--foreground-rgb));
-    background: linear-gradient(to bottom,
-        transparent,
-        rgb(var(--background-end-rgb))) rgb(var(--background-start-rgb));
-
     min-height: 100vh;
   }
 }
